third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGLength-px-expected.txt

This is a testharness.js-based test.
[FAIL] SVGLength, converting from 'px' to other units (detached), percentage
  Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Could not resolve relative length.
[FAIL] SVGLength, converting from 'px' to other units (detached), ems
  Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Could not resolve relative length.
[FAIL] SVGLength, converting from 'px' to other units (detached), exs
  Failed to execute 'convertToSpecifiedUnits' on 'SVGLength': Could not resolve relative length.
[FAIL] SVGLength, converting from 'px' to other units (detached), q
  assert_approx_equals: expected 45.3438 +/- 0.01 but got 45.354331970214844
Harness: the test ran to completion.
